[VACGOM-69] refactor: Id Long -> UUID로 변경
