+++ src/main/kotlin/kr/co/vacgom/api/auth/security/UserAuthentication.kt
@@ -5,7 +5,7 @@ import org.springframework.security.core.GrantedAuthority
 import java.util.*
 
 class UserAuthentication(
-    private val userId: Long,
+    private val userId: UUID,
     authorities: Collection<GrantedAuthority>,
 ) : AbstractAuthenticationToken(authorities) {
