+++ src/main/kotlin/kr/co/vacgom/api/baby/domain/enums/Gender.kt
@@ -1,0 +1,5 @@
+package kr.co.vacgom.api.baby.domain.enums
+
+enum class Gender {
+    MALE, FEMALE,
+}
