+++ src/main/kotlin/kr/co/vacgom/api/baby/repository/BabyJpaRepository.kt
@@ -2,5 +2,6 @@ package kr.co.vacgom.api.baby.repository
 
 import kr.co.vacgom.api.baby.domain.Baby
 import org.springframework.data.jpa.repository.JpaRepository
+import java.util.UUID
 
-interface BabyJpaRepository: JpaRepository<Baby, Long>
+interface BabyJpaRepository: JpaRepository<Baby, UUID>
