+++ src/main/kotlin/kr/co/vacgom/api/user/repository/UserJpaRepository.kt
@@ -2,7 +2,8 @@ package kr.co.vacgom.api.user.repository
 
 import kr.co.vacgom.api.user.domain.User
 import org.springframework.data.jpa.repository.JpaRepository
+import java.util.*
 
-interface UserJpaRepository : JpaRepository<User, Long> {
+interface UserJpaRepository : JpaRepository<User, UUID> {
     fun findBySocialId(socialId: String) : User?
 }
