+++ src/main/kotlin/kr/co/vacgom/api/user/repository/UserRepository.kt
@@ -1,11 +1,12 @@
 package kr.co.vacgom.api.user.repository
 
 import kr.co.vacgom.api.user.domain.User
+import java.util.*
 
 
 interface UserRepository {
     fun save(user: User): User
     fun findBySocialId(socialId: String): User?
-    fun findById(userId: Long): User?
-    fun deleteById(userId: Long)
+    fun findById(userId: UUID): User?
+    fun deleteById(userId: UUID)
 }
