+++ src/main/kotlin/kr/co/vacgom/api/user/repository/UserRepositoryAdapter.kt
@@ -3,6 +3,7 @@ package kr.co.vacgom.api.user.repository
 import kr.co.vacgom.api.user.domain.User
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Repository
+import java.util.*
 
 @Repository
 class UserRepositoryAdapter(
@@ -16,11 +17,11 @@ class UserRepositoryAdapter(
         return userJpaRepository.findBySocialId(socialId)
     }
 
-    override fun findById(userId: Long): User? {
+    override fun findById(userId: UUID): User? {
         return userJpaRepository.findByIdOrNull(userId)
     }
 
-    override fun deleteById(userId: Long) {
+    override fun deleteById(userId: UUID) {
         userJpaRepository.deleteById(userId)
     }
 }
